fixes warning
about sign of unsigned result still unsigned (MSVC)

+++ src/game/lara.c
@@ -2150,7 +2150,7 @@ int32_t TestLaraVault(ITEM_INFO *item, COLL_INFO *coll)
         item->anim_number = AA_STOP;
         item->frame_number = AF_STOP * ANIM_SCALE;
         Lara.calc_fall_speed =
-            -(phd_sqrt((int)(-2 * GRAVITY * (hdif + 800))) + 3);
+            -(int16_t)(phd_sqrt((int)(-2 * GRAVITY * (hdif + 800))) + 3);
         AnimateLara(item);
         item->pos.y_rot = angle;
         ShiftItemLara(item, coll);
